pyproject.toml
[project]
name = "hello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown>=3.7",
    "python-fasthtml>=0.12.1",
]
